auth updated and refined
