+++ package.json
@@ -42,18 +42,18 @@
     "@apollo/experimental-nextjs-app-support": "^0.8.0",
     "@headlessui/react": "^1.7.18",
     "@heroicons/react": "^2.1.1",
-    "@next/bundle-analyzer": "^13.5.6",
-    "@next/env": "^13.5.6",
+    "@next/bundle-analyzer": "^14.1.3",
+    "@next/env": "^14.1.3",
     "@sentry/nextjs": "^7.106.1",
-    "@takeshape/next-auth-all-access": "^1.0.0",
+    "@takeshape/next-auth-all-access": "^1.1.0-5",
     "@takeshape/routing": "^10.7.6",
     "cookies-next": "^4.1.1",
     "date-fns": "^3.4.0",
     "graphql": "^16.8.1",
     "imask": "^7.5.0",
     "jotai": "^2.7.0",
-    "next": "^13.5.6",
-    "next-auth": "^4.24.7",
+    "next": "^14.1.3",
+    "next-auth": "^5.0.0-beta.15",
     "next-recaptcha-v3": "^1.3.0",
     "next-seo": "^6.5.0",
     "nookies": "^2.5.2",
