+++ src/app/api/auth-all-access/[...allaccess]/route.ts
@@ -1,0 +1,1 @@
+export { GET } from '@/auth-all-access';
