+++ src/app/checkout/page.tsx
@@ -1,9 +1,9 @@
+import { auth } from '@/auth';
 import { AuthProvider } from '@/features/Auth/AuthProvider';
 import { CheckoutAfterSignIn } from '@/features/Cart/CheckoutAfterSignIn';
-import { getServerSession } from 'next-auth';
 
 export default async function CheckoutPage() {
-  const session = await getServerSession();
+  const session = await auth();
   return (
     <AuthProvider session={session}>
       <CheckoutAfterSignIn />
